Move plotting functionality out, compare_benches no longer owns plot-related data
Introduced _nvbench_compare_plotting.py helper that contains class
PlotCollector.

PlotCollector owns plotting-related state and rendering, and compare_benches()
interacts with plotting through that collector.

compare_benches() still produces comparison facts, but delegates plotting state
management and rendering to PlotCollector.
